# openapi.yaml
openapi: 3.0.3
info:
  title: ShopBuilder API
  version: 1.0.0
  description: >
    REST API contract for ShopBuilder, a multi-tenant Shopify-like e-commerce backend
    built with FastAPI, SQLModel, PostgreSQL, and Redis.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new store owner and create the first store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Store created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
              examples:
                created:
                  value:
                    access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    refresh_token: "d4ce0cfb-62ac-4d75-9dbe-69069485d40f"
                    token_type: "bearer"
                    expires_in: 900
                    store:
                      id: "store_01HZX8D5P7J9M3R4K5L6N7Q8R"
                      name: "Nomad Home"
                      subdomain: "nomad-home"
                      status: "draft"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate a user and issue JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a valid refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke a refresh token and log out
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Refresh token revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/email/verify:
    post:
      tags: [Auth]
      summary: Verify an email address with a one-time token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationRequest'
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/email/resend:
    post:
      tags: [Auth]
      summary: Send a new email verification token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
      responses:
        '200':
          description: Verification email queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/password/forgot:
    post:
      tags: [Auth]
      summary: Queue a password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset email queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/password/reset:
    post:
      tags: [Auth]
      summary: Reset password using a one-time token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /stores/{store_id}/products:
    get:
      tags: [Products]
      summary: List products for a store
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StoreIdParam'
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, active, archived]
        - in: query
          name: sort
          schema:
            type: string
            example: "-updated_at"
      responses:
        '200':
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Products]
      summary: Create a product with options and variant generation rules
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StoreIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /stores/{store_id}/webhooks/events:
    get:
      tags: [Webhooks]
      summary: List inbound webhook events for a store
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StoreIdParam'
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: status
          schema:
            type: string
            enum: [queued, processing, processed, failed, dead_letter]
        - in: query
          name: sort
          schema:
            type: string
            example: "-received_at"
      responses:
        '200':
          description: Paginated webhook event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEventListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /stores/{store_id}/webhooks/inbound:
    post:
      tags: [Webhooks]
      summary: Receive a signed inbound webhook event
      parameters:
        - $ref: '#/components/parameters/StoreIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookInboundRequest'
      responses:
        '202':
          description: Event persisted for asynchronous processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CursorParam:
      in: query
      name: cursor
      required: false
      schema:
        type: string
        example: "eyJzb3J0X2tleSI6IjIwMjYtMDQtMjRUMTA6MzA6MDBaIiwiaWQiOiJvcmRfMDEifQ=="
      description: Opaque cursor for keyset pagination.
    LimitParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    StoreIdParam:
      in: path
      name: store_id
      required: true
      schema:
        type: string
        example: "store_01HZX8D5P7J9M3R4K5L6N7Q8R"
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "bad_request"
              message: "Request payload is malformed."
              request_id: "req_01HZYA9RBK6P2Q11YTZ6K9M7K3"
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "unauthorized"
              message: "Missing or invalid bearer token."
              request_id: "req_01HZYA9RBK6P2Q11YTZ6K9M7K3"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "not_found"
              message: "Requested store was not found."
              request_id: "req_01HZYA9RBK6P2Q11YTZ6K9M7K3"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "conflict"
              message: "Subdomain is already reserved."
              request_id: "req_01HZYA9RBK6P2Q11YTZ6K9M7K3"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "internal_error"
              message: "Unexpected server error."
              request_id: "req_01HZYA9RBK6P2Q11YTZ6K9M7K3"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
    ValidationError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id, fields]
          properties:
            code:
              type: string
              example: "validation_error"
            message:
              type: string
              example: "One or more fields are invalid."
            request_id:
              type: string
              example: "req_01HZYA9RBK6P2Q11YTZ6K9M7K3"
            fields:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                    example: "email"
                  message:
                    type: string
                    example: "Value must be a valid email address."
    AuthTokens:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          example: 900
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "owner@nomad-home.kz"
        password:
          type: string
          minLength: 8
          example: "StrongPass!2026"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "owner@nomad-home.kz"
        password:
          type: string
          example: "StrongPass!2026"
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          example: "c29tZS1yZWZyZXNoLXRva2Vu"
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          example: "c29tZS1yZWZyZXNoLXRva2Vu"
    EmailVerificationRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          example: "secure-email-verification-token"
    ResendVerificationRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: "owner@nomad-home.kz"
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: "owner@nomad-home.kz"
    ResetPasswordRequest:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
          example: "secure-password-reset-token"
        new_password:
          type: string
          minLength: 8
          example: "NewStrongPass!2026"
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: "Email address verified."
...
    PaginationMeta:
      type: object
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: string
          nullable: true
          example: "eyJwbGFjZWRfYXQiOiIyMDI2LTA0LTI0VDEwOjMwOjAwWiIsImlkIjoib3JkXzAxIn0="
        has_more:
          type: boolean
          example: true
...
    ProductOptionRequest:
      type: object
      required: [name, values]
      properties:
        name:
          type: string
          example: "Size"
        values:
          type: array
          items:
            type: string
          example: ["S", "M", "L"]
    ProductCreateRequest:
      type: object
      required: [title, handle, status, options]
      properties:
        title:
          type: string
          example: "Minimal Hoodie"
        handle:
          type: string
          example: "minimal-hoodie"
        status:
          type: string
          enum: [draft, active, archived]
          example: "draft"
        description_html:
          type: string
          example: "<p>Heavyweight hoodie made for everyday wear.</p>"
        vendor:
          type: string
          example: "Nomad Essentials"
        options:
          type: array
          items:
            $ref: '#/components/schemas/ProductOptionRequest'
    Variant:
      type: object
      required: [id, sku, title, price_amount, is_active]
      properties:
        id:
          type: string
          example: "var_01HZX9C8Y0CFV0A6D4DVV18YQ2"
        sku:
          type: string
          example: "NH-HOODIE-BLK-M"
        title:
          type: string
          example: "Black / M"
        price_amount:
          type: integer
          example: 18990
        compare_at_amount:
          type: integer
          nullable: true
          example: 22990
        is_active:
          type: boolean
          example: true
    Product:
      type: object
      required: [id, title, handle, status, variants]
      properties:
        id:
          type: string
          example: "prod_01HZX92TH06CS4T0D8K0X1F3PW"
        title:
          type: string
          example: "Minimal Hoodie"
        handle:
          type: string
          example: "minimal-hoodie"
        status:
          type: string
          enum: [draft, active, archived]
        vendor:
          type: string
          example: "Nomad Essentials"
        variants:
          type: array
          items:
            $ref: '#/components/schemas/Variant'
    ProductListResponse:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        page:
          $ref: '#/components/schemas/PaginationMeta'
...
    WebhookInboundRequest:
      type: object
      required: [provider_name, provider_event_id, event_type, occurred_at, payload, signature]
      properties:
        provider_name:
          type: string
          example: "kaspi_pay"
        provider_event_id:
          type: string
          example: "evt_9c2dbfe8-6d34-4823-81a5-46e0f9c2287a"
        event_type:
          type: string
          example: "payment.succeeded"
        occurred_at:
          type: string
          format: date-time
          example: "2026-04-24T10:33:00Z"
        signature:
          type: string
          example: "t=1713954780,v1=3e9adfe13e2d75b61634c4f325f6c0385fb75767f7be52bbd432f07745d16470"
        payload:
          type: object
          additionalProperties: true
          example:
            order_id: "ord_01HZXC3M92J6D7V7N0P6R51QJ4"
            amount: 40480
            currency: "KZT"
            status: "captured"
    WebhookAccepted:
      type: object
      required: [event_id, status]
      properties:
        event_id:
          type: string
          example: "wh_01HZXD9GBAZ93VTBK5Q3R8WQ4C"
        status:
          type: string
          example: "queued"
    WebhookEvent:
      type: object
      required: [id, provider_name, provider_event_id, event_type, status, attempt_count, received_at]
      properties:
        id:
          type: string
          example: "wh_01HZXD9GBAZ93VTBK5Q3R8WQ4C"
        provider_name:
          type: string
          example: "kaspi_pay"
        provider_event_id:
          type: string
          example: "evt_9c2dbfe8-6d34-4823-81a5-46e0f9c2287a"
        event_type:
          type: string
          example: "payment.succeeded"
        status:
          type: string
          enum: [queued, processing, processed, failed, dead_letter]
        attempt_count:
          type: integer
          example: 1
        last_error:
          type: string
          nullable: true
          example: null
        received_at:
          type: string
          format: date-time
          example: "2026-04-24T10:33:01Z"
    WebhookEventListResponse:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEvent'
        page:
          $ref: '#/components/schemas/PaginationMeta'